+++ samples/Sample_06_Footnote.php
@@ -7,7 +7,7 @@ $phpWord = new \PhpOffice\PhpWord\PhpWord();
 \PhpOffice\PhpWord\Settings::setCompatibility(false);
 
 // New portrait section
-$section = $phpWord->createSection();
+$section = $phpWord->addSection();
 
 // Add style definitions
 $phpWord->addParagraphStyle('pStyle', array('spacing'=>100));
@@ -16,10 +16,10 @@ $phpWord->addFontStyle('ColoredText', array('color'=>'FF8080'));
 $phpWord->addLinkStyle('NLink', array('color'=>'0000FF', 'underline' => \PhpOffice\PhpWord\Style\Font::UNDERLINE_SINGLE));
 
 // Add text elements
-$textrun = $section->createTextRun('pStyle');
+$textrun = $section->addTextRun('pStyle');
 $textrun->addText('This is some lead text in a paragraph with a following footnote. ','pStyle');
 
-$footnote = $textrun->createFootnote();
+$footnote = $textrun->addFootnote();
 $footnote->addText('Just like a textrun, a footnote can contain native texts. ');
 $footnote->addText('No break is placed after adding an element. ', 'BoldText');
 $footnote->addText('All elements are placed inside a paragraph. ', 'ColoredText');
@@ -34,7 +34,7 @@ $footnote->addObject('resources/_sheet.xls');
 $footnote->addText('But you can only put footnote in section, not in header or footer.');
 
 $section->addText('You can also create the footnote directly from the section making it wrap in a paragraph like the footnote below this paragraph. But is is best used from within a textrun.');
-$footnote = $section->createFootnote();
+$footnote = $section->addFootnote();
 $footnote->addText('The reference for this is wrapped in its own line');
 
 // Save file