+++ samples/Sample_12_HeaderFooter.php
@@ -6,10 +6,10 @@ echo date('H:i:s') , " Create new PhpWord object" , \EOL;
 $phpWord = new \PhpOffice\PhpWord\PhpWord();
 
 // New portrait section
-$section = $phpWord->createSection();
+$section = $phpWord->addSection();
 
 // Add first page header
-$header = $section->createHeader();
+$header = $section->addHeader();
 $header->firstPage();
 $table = $header->addTable();
 $table->addRow();
@@ -23,11 +23,11 @@ $table->addCell(4500)->addImage(
 );
 
 // Add header for all other pages
-$subsequent = $section->createHeader();
+$subsequent = $section->addHeader();
 $subsequent->addText("Subsequent pages in Section 1 will Have this!");
 
 // Add footer
-$footer = $section->createFooter();
+$footer = $section->addFooter();
 $footer->addPreserveText('Page {PAGE} of {NUMPAGES}.', array('align' => 'center'));
 $footer->addLink('http://google.com', 'Direct Google');
 
@@ -50,9 +50,9 @@ $section->addTextBreak();
 $section->addText('Some text...');
 
 // New portrait section
-$section2 = $phpWord->createSection();
+$section2 = $phpWord->addSection();
 
-$sec2Header = $section2->createHeader();
+$sec2Header = $section2->addHeader();
 $sec2Header->addText("All pages in Section 2 will Have this!");
 
 // Write some text