+++ samples/Sample_18_Watermark.php
@@ -7,8 +7,8 @@ $phpWord = new \PhpOffice\PhpWord\PhpWord();
 
 // Begin code
 
-$section = $phpWord->createSection();
-$header = $section->createHeader();
+$section = $phpWord->addSection();
+$header = $section->addHeader();
 $header->addWatermark('resources/_earth.jpg', array('marginTop' => 200, 'marginLeft' => 55));
 $section->addText('The header reference to the current section includes a watermark image.');