+++ samples/Sample_20_BGColor.php
@@ -4,7 +4,7 @@ include_once 'Sample_Header.php';
 // New Word document
 echo date('H:i:s'), " Create new PhpWord object", \EOL;
 $phpWord = new \PhpOffice\PhpWord\PhpWord();
-$section = $phpWord->createSection();
+$section = $phpWord->addSection();
 
 $section->addText("This is some text highlighted using fgColor (limited to 15 colors)     ", array("fgColor" => \PhpOffice\PhpWord\Style\Font::FGCOLOR_YELLOW));
 $section->addText("This one uses bgColor and is using hex value (0xfbbb10)", array("bgColor" => "fbbb10"));