+++ samples/Sample_22_CheckBox.php
@@ -5,7 +5,7 @@ include_once 'Sample_Header.php';
 echo date('H:i:s'), " Create new PhpWord object", \EOL;
 $phpWord = new \PhpOffice\PhpWord\PhpWord();
 
-$section = $phpWord->createSection();
+$section = $phpWord->addSection();
 $section->addText('Check box in section');
 $section->addCheckBox('chkBox1', 'Checkbox 1');
 $section->addText('Check box in table cell');