+++ tests/PhpWord/Tests/Style/ParagraphTest.php
@@ -105,7 +105,7 @@ class ParagraphTest extends \PHPUnit_Framework_TestCase
     public function testLineHeight()
     {
         $phpWord = new PhpWord();
-        $section = $phpWord->createSection();
+        $section = $phpWord->addSection();
 
         // Test style array
         $text = $section->addText('This is a test', array(), array(