+++ tests/PhpWord/Tests/Writer/RTFTest.php
@@ -52,7 +52,7 @@ class RTFTest extends \PHPUnit_Framework_TestCase
         $phpWord = new PhpWord();
         $phpWord->addFontStyle('Font', array('name' => 'Verdana', 'size' => 11, 'color' => 'FF0000', 'fgColor' => 'FF0000'));
         $phpWord->addParagraphStyle('Paragraph', array('align' => 'center'));
-        $section = $phpWord->createSection();
+        $section = $phpWord->addSection();
         $section->addText('Test 1', 'Font', 'Paragraph');
         $section->addTextBreak();
         $section->addText('Test 2', array('name' => 'Tahoma', 'bold' => true, 'italic' => true));
@@ -64,8 +64,8 @@ class RTFTest extends \PHPUnit_Framework_TestCase
         $section->addImage($imageSrc);
         $section->addObject($objectSrc);
         $section->addTOC();
-        $section = $phpWord->createSection();
-        $textrun = $section->createTextRun();
+        $section = $phpWord->addSection();
+        $textrun = $section->addTextRun();
         $textrun->addText('Test 3');
         $textrun->addTextBreak();
         $writer = new RTF($phpWord);
@@ -84,7 +84,7 @@ class RTFTest extends \PHPUnit_Framework_TestCase
     public function testSavePhpOutput()
     {
         $phpWord = new PhpWord();
-        $section = $phpWord->createSection();
+        $section = $phpWord->addSection();
         $section->addText('Test');
         $writer = new RTF($phpWord);
         $writer->save('php://output');